requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.0
matplotlib==3.9.2
mpmath==1.3.0
ncnn==1.0.20240820
networkx==3.3
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
psutil==6.0.0
py-cpuinfo==9.0.0
pydobot==1.3.2
pydobotplus==0.1.2
pyparsing==3.1.4
pyserial==3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
RPi.GPIO==0.7.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.7
ultralytics-thop==2.0.9
urllib3==2.2.3
